Split pool manager
